fix edit form

+++ server/routes/users.js
@@ -28,12 +28,22 @@ export default (app) => {
 
       return reply;
     })
-    .get('/users/:id/edit', { name: 'editUser' }, async (req, reply) => {
+    .get('/users/:id/edit', { name: 'editUser', preValidation: app.authenticate }, async (req, reply) => {
+      if (req.user.id !== Number(req.params.id)) {
+        req.flash('error', i18next.t('flash.accessDenied'));
+        reply.redirect(app.reverse('root'));
+        return reply;
+      }
       const user = await app.objection.models.user.query().findById(req.params.id);
       reply.render('users/edit', { user });
       return reply;
     })
-    .patch('/users/:id', async (req, reply) => {
+    .patch('/users/:id', { preValidation: app.authenticate }, async (req, reply) => {
+      if (req.user.id !== Number(req.params.id)) {
+        req.flash('error', i18next.t('flash.accessDenied'));
+        reply.redirect(app.reverse('root'));
+        return reply;
+      }
       const user = await app.objection.models.user.query().findById(req.params.id);
 
       try {
@@ -47,7 +57,12 @@ export default (app) => {
 
       return reply;
     })
-    .delete('/users/:id', async (req, reply) => {
+    .delete('/users/:id', { name: 'user', preValidation: app.authenticate }, async (req, reply) => {
+      if (req.user.id !== Number(req.params.id)) {
+        req.flash('error', i18next.t('flash.accessDenied'));
+        reply.redirect(app.reverse('root'));
+        return reply;
+      }
       await app.objection.models.user.query().deleteById(req.params.id);
       req.flash('info', i18next.t('flash.users.delete.success'));
       reply.redirect(app.reverse('users'));
